Move Tests file to its own directory
